moved casacore stuff into casacore util file
