+++ CMakeLists.txt
@@ -250,6 +250,7 @@ set(SOURCE_FILES
         src/OnMessageTask.cc
         src/FileSettings.cc
         src/Util.cc
+        src/Util/Casacore.cc
         src/Util/File.cc
         src/Util/Image.cc
         src/Util/Message.cc
