+++ src/Util/File.cc
@@ -30,50 +30,6 @@ bool IsCompressedFits(const std::string& filename) {
     return is_fits;
 }
 
-void GetSpectralCoordPreferences(
-    casacore::ImageInterface<float>* image, bool& prefer_velocity, bool& optical_velocity, bool& prefer_wavelength, bool& air_wavelength) {
-    prefer_velocity = optical_velocity = prefer_wavelength = air_wavelength = false;
-    casacore::CoordinateSystem coord_sys(image->coordinates());
-    if (coord_sys.hasSpectralAxis()) { // prefer spectral axis native type
-        casacore::SpectralCoordinate::SpecType native_type;
-        if (image->imageType() == "CartaMiriadImage") { // workaround to get correct native type
-            carta::CartaMiriadImage* miriad_image = static_cast<carta::CartaMiriadImage*>(image);
-            native_type = miriad_image->NativeType();
-        } else {
-            native_type = coord_sys.spectralCoordinate().nativeType();
-        }
-        switch (native_type) {
-            case casacore::SpectralCoordinate::FREQ: {
-                break;
-            }
-            case casacore::SpectralCoordinate::VRAD:
-            case casacore::SpectralCoordinate::BETA: {
-                prefer_velocity = true;
-                break;
-            }
-            case casacore::SpectralCoordinate::VOPT: {
-                prefer_velocity = true;
-
-                // Check doppler type; oddly, native type can be VOPT but doppler is RADIO--?
-                casacore::MDoppler::Types vel_doppler(coord_sys.spectralCoordinate().velocityDoppler());
-                if ((vel_doppler == casacore::MDoppler::Z) || (vel_doppler == casacore::MDoppler::OPTICAL)) {
-                    optical_velocity = true;
-                }
-                break;
-            }
-            case casacore::SpectralCoordinate::WAVE: {
-                prefer_wavelength = true;
-                break;
-            }
-            case casacore::SpectralCoordinate::AWAV: {
-                prefer_wavelength = true;
-                air_wavelength = true;
-                break;
-            }
-        }
-    }
-}
-
 int GetStokesValue(const CARTA::StokesType& stokes_type) {
     int stokes_value(-1);
     switch (stokes_type) {
